# Spend less time waiting for LocalTaskJob's subprocss process to finish