--- a/airflow/task/task_runner/standard_task_runner.py
+++ b/airflow/task/task_runner/standard_task_runner.py
@@ -108,7 +108,10 @@ class StandardTaskRunner(BaseTaskRunner):
         if self.process is None:
             return
 
-        if self.process.is_running():
+        # Reap the child process - it may already be finished
+        _ = self.return_code(timeout=0)
+
+        if self.process and self.process.is_running():
             rcs = reap_process_group(self.process.pid, self.log)
             self._rc = rcs.get(self.process.pid)