--- a/tests/dags/test_heartbeat_failed_fast.py
+++ b/tests/dags/test_heartbeat_failed_fast.py
@@ -30,5 +30,5 @@ args = {
 dag = DAG(dag_id='test_heartbeat_failed_fast', default_args=args)
 task = BashOperator(
     task_id='test_heartbeat_failed_fast_op',
-    bash_command='sleep 5',
+    bash_command='sleep 7',
     dag=dag)