@potiuk 2020-10-12
> One heisen test failed in a slightly odd way:
> 
> ```
> self = <airflow.executors.dask_executor.DaskExecutor object at 0x7efe33cd1090>
> 
>     def end(self) -> None:
>         if not self.client:
>             raise AirflowException(NOT_STARTED_MESSAGE)
>         if not self.futures:
> >           raise AirflowException(NOT_STARTED_MESSAGE)
> E           airflow.exceptions.AirflowException: The executor should be started first!
> 
> airflow/executors/dask_executor.py:113: AirflowException
> ```
> 
> Do we think it's safe to merge anyway?

You can run the tests easily now for the whole group :) . LEt me do it for you