@potiuk 2020-10-12
It failed on me with the first attempt I made. So the problem is real.

```
./breeze --github-image-id 302320596 --backend mysql --mysql-version 5.7 --python 3.8 --db-reset --test-type Core tests
```

![Screenshot from 2020-10-12 20-34-17](https://user-images.githubusercontent.com/595491/95779459-67d6f580-0cca-11eb-87e7-5e69426fe40a.png)