@ashb 2020-10-12
I have a fix for the Dask Executor, but I'm not sure what is going really on (why it is suddenly failing.)

It has

```
if not self.futures:
```

But the problem is that `self.futures` is a dict, so evaluates to false when it's empty.